Modified make file and divied pthread file
Modified make file and divied pthread file.
Delete don't used file.
