+++ .gitignore
@@ -3,3 +3,4 @@
 *.exe
 *.cson
 *.json
+*.o
