+++ Makefile
@@ -3,12 +3,19 @@
 CC = gcc
 CFLAGS = -W
 LIB = -lpthread
+OBJ = main.o pthread_main.o
+RM = del
+#RM = rm -f
 
-hello : main.o
-    $(CC) $(CFLAGS) -o hello.exe main.o -g $(LIB)
+hello : $(OBJ)
+    $(CC) $(CFLAGS) -o hello.exe $(OBJ) -g  $(LIB)
 
 main.o : main.c
     $(CC) $(CFLAGS) -c main.c -g
 	
+pthread_main.o : pthread_main.c
+    $(CC) $(CFLAGS) -c pthread_main.c -g
+
 clean :
-	rm hello.exe
+    $(RM) hello.exe $(OBJ)
+    echo "hello.exe is deleted"
