--- main0.c
@@ -1,6 +1,0 @@
-#include<stdio.h>
-
-int main( int argc, char *argv[])
-{
-    printf("Hello world!\n");
-}
